requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.120.3
h11==0.16.0
idna==3.11
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.49.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
wheel==0.45.1
